# docs/swagger.yaml
definitions:
  domain.Chime:
    properties:
      author:
        type: string
      chimeContent:
        type: string
      chimeTitle:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isPrivate:
        type: boolean
    type: object
  domain.ChimeCreateRequest:
    properties:
      chime_content:
        type: string
      chime_title:
        type: string
      is_private:
        type: boolean
    type: object
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  domain.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  domain.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      penName:
        type: string
      userName:
        type: string
    required:
    - email
    - password
    - userName
    type: object
  domain.SignUpResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.User:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      penName:
        type: string
      userName:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login/:
    post:
      consumes:
      - application/json
      description: Logs a user in by validating credentials and returning access and
        refresh tokens.
      parameters:
      - description: Login Request Payload
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful, returns access and refresh tokens
          schema:
            $ref: '#/definitions/domain.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized - Invalid credentials or user not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /auth/signup/:
    post:
      consumes:
      - application/json
      description: Signs up a new user by validating the fields and returns their
        access token and refresh token
      parameters:
      - description: SignUp Request Payload
        in: body
        name: signUpRequest
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Signup successful, returns access and refresh tokens
          schema:
            $ref: '#/definitions/domain.SignUpResponse'
        "400":
          description: Bad Request - Invalid or missing parameters in the request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Conflict - User already exists
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error - Issue processing the request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: User Signup
      tags:
      - Authentication
  /chimes/:
    get:
      consumes:
      - application/json
      description: Fetch all Chimes from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Chime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all Chimes
      tags:
      - Chimes
    post:
      consumes:
      - application/json
      description: Create a new chime by providing a request body with necessary details.
      parameters:
      - description: Chime Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChimeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Chime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Creates a new Chime
      tags:
      - Chimes
  /chimes/user/:
    get:
      consumes:
      - application/json
      description: Fetch all Chimes from logged in user from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Chime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all Chimes of logged in user
      tags:
      - Chimes
securityDefinitions:
  BearerAuth:
    description: API documentation for ChimeSpace backend
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"